channel.ts created to separate types from release
