+++ release-calendar/src/types/channel.ts
@@ -1,0 +1,10 @@
+export enum Channel {
+  LTS = 'lts',
+  STABLE = 'stable',
+  BETA_ONE_PERCENT = 'betaOnePercent',
+  EXPERIMENTAL_ONE_PERCENT = 'experimentalOnePercent',
+  BETA_OPTIN = 'betaOptin',
+  EXPERIMENTAL_OPTIN = 'experimentalOptin',
+  NIGHTLY = 'nightly',
+  ERROR = 'error',
+}
